Re-worked super-class call
Script nodes can now call 'self.super(<ClassType>)' to get an object that acts as
an wrapper to a super object, just like the 'super' built-in.

We need to use a special method due to the way the AVANGO nodes are
wrapped.

This is now the preferred way to call methods of base classes, in
particular the '__init__' method.

Example:
  class A(avango.script.Script):
    def __init__(self):
      self.super(A).__init__()
      self.value = 42

As a side note, call the base '__init__' method of classes when
implementing a '__init__' method yourself in classes derived from
avango.script.Script. This will be mandatory in the future.